> Refactored `$SE.hsv(...)` to `$Graphics.hsv(...)`
