+++ res/program
@@ -1,5 +1,5 @@
 name:{Stipple Effect}
-version:{1.2.3.29}
+version:{1.2.3.34}
 devbuild:{true}
 native_standard:{1.3}
 palette_standard:{1.0}
